增加修改密码功能

+++ www/php/delete.php
@@ -2,8 +2,13 @@
     include('connect.php');
     
     $addr=$_POST['address'];
-    $user=$_SESSION['username'];
-    
+    $username=$_SESSION["username"];
+    $sql="SELECT from image where address='$addr'&&username='$username'";
+    if($sql){
+        echo"<script>alert('你无权删除此文件');</script>";
+        header('Location: ../index.html'); //跳转登录页面
+        exit();
+    }
     $sql="DELETE from image where address='$addr'";
     mysqli_query($con, $sql);
     
@@ -15,3 +20,5 @@
     }
     
     
+
+
